requirements.txt
plotly==4.14.3
pandas==1.2.0
streamlit==0.81.1
streamlit_plotly_events==0.0.6
numpy==1.18.5
Pillow==8.2.0
scikit_learn==0.24.2
